Consolidate example projects by dependency
