+++ examples-build/Examples.Core/Examples.Core.csproj
@@ -9,6 +9,7 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="10.0.1" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.0.1" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="10.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="10.0.1" />
   </ItemGroup>
 
   <ItemGroup>
@@ -17,16 +18,18 @@
 
   <!--
     The example .cs files live under modules/<m>/examples (Antora needs them there).
-    Non-recursive globs pick up the loose snippet files without descending into the
-    standalone-project subdirs (Couchbase.Examples.*). Encryption/Transactions files
-    are excluded here because they build in their own projects. devguide is listed
+    Non-recursive globs pick up every loose snippet in each examples root: this now
+    includes the main-SDK Transactions, Search, RangeScan and SubDocument examples.
+    The non-recursive glob deliberately does NOT descend into howtos/examples/logging
+    (built by the Examples.Logging project). EncryptingUsingSdk.cs is excluded because
+    it needs the separate encryption package (Examples.Encryption). devguide is listed
     explicitly because most of its files are orphaned and out of scope.
   -->
   <ItemGroup>
     <Compile Include="../../modules/hello-world/examples/*.cs" />
     <Compile Include="../../modules/concept-docs/examples/*.cs" />
     <Compile Include="../../modules/howtos/examples/*.cs"
-             Exclude="../../modules/howtos/examples/TransactionsExample.cs;../../modules/howtos/examples/EncryptingUsingSdk.cs" />
+             Exclude="../../modules/howtos/examples/EncryptingUsingSdk.cs" />
     <Compile Include="../../modules/devguide/examples/dotnet/Analytics.cs" />
     <Compile Include="../../modules/devguide/examples/dotnet/ConnectionBase.cs" />
     <Compile Include="../../modules/devguide/examples/dotnet/Cloud/Cloud.cs" />
