+++ modules/devguide/examples/dotnet/Analytics.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
 namespace Couchbase.Net.DevGuide
 {
