--- modules/howtos/examples/Couchbase.Examples.KV/SubDocument/SubDocument.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Library</OutputType>
-        <TargetFramework>net10.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="CouchbaseNetClient" Version="3.9.2" />
-    </ItemGroup>
-
-</Project>
