--- modules/howtos/examples/Couchbase.Examples.Logging.GenericHost/appsettings.json
@@ -1,7 +1,0 @@
-﻿{
-  "couchbase": {
-    "connectionString": "couchbase://your-ip",
-    "username": "Administrator",
-    "password": "password"
-  }
-}
