+++ modules/howtos/examples/EncryptingUsingSdk.cs
@@ -1,9 +1,5 @@
-using System;
 using System.Security.Cryptography;
-using System.Security.Cryptography.X509Certificates;
 using System.Text;
-using System.Threading.Tasks;
-using Couchbase;
 using Couchbase.Encryption;
 using Couchbase.Encryption.Attributes;
 using Couchbase.Encryption.Internal;
