+++ modules/howtos/examples/logging/GenericHost.cs
