+++ modules/howtos/examples/logging/NoHost.cs
