+++ modules/howtos/examples/logging/Worker.cs
