+++ modules/howtos/examples/logging/windows/App.config
