+++ modules/howtos/examples/logging/windows/Log4NetExtensions.cs
