+++ modules/howtos/examples/logging/windows/Log4NetLogger.cs
