+++ modules/howtos/examples/logging/windows/Log4NetProvider.cs
