+++ modules/howtos/examples/logging/windows/Program.cs
