+++ modules/howtos/examples/logging/windows/Properties/AssemblyInfo.cs
