+++ modules/howtos/examples/logging/windows/log4net.config
