+++ modules/howtos/examples/logging/windows/packages.config
