+++ modules/howtos/pages/kv-operations.adoc
@@ -270,7 +270,7 @@ Here's an example of a KV range scan that gets all documents in a collection:
 .KV Range Scan for all documents in a collection
 [source,csharp]
 ----
-include::example$Couchbase.Examples.KV/RangeScan/RangeScan/Program.cs[tag=rangeScanAllDocuments,indent=0]
+include::example$RangeScan.cs[tag=rangeScanAllDocuments,indent=0]
 ----
 <1> The `RangeScan()` constructor has two optional nullable parameters: `from` and `to`.
 If you pass null like in this example, you'll get all documents in the collection.
@@ -291,7 +291,7 @@ For example, to get all documents associated with user "alice", you would write:
 .KV Range Scan for all documents in a collection whose IDs start with "alice::"
 [source,csharp]
 ----
-include::example$Couchbase.Examples.KV/RangeScan/RangeScan/Program.cs[tag=rangeScanPrefix,indent=0]
+include::example$RangeScan.cs[tag=rangeScanPrefix,indent=0]
 ----
 <1> Note the scan type is *prefixScan*.
 
@@ -303,7 +303,7 @@ If you want to get random documents from a collection, use a sample scan.
 .KV Range Scan for 100 random documents
 [source,csharp]
 ----
-include::example$Couchbase.Examples.KV/RangeScan/RangeScan/Program.cs[tag=rangeScanSample,indent=0]
+include::example$RangeScan.cs[tag=rangeScanSample,indent=0]
 ----
 <1> In this example, no more than `100` documents are returned.
 
@@ -315,7 +315,7 @@ If you only want the document IDs, set the `idsOnly` option to true, like this:
 .KV Range Scan for all document IDs in a collection
 [source,csharp]
 ----
-include::example$Couchbase.Examples.KV/RangeScan/RangeScan/Program.cs[tag=rangeScanAllDocumentIds,indent=0]
+include::example$RangeScan.cs[tag=rangeScanAllDocumentIds,indent=0]
 ----
 <1> The returned `ScanResult` objects throw `NoSuchElementException` if you try to access any property other than `Id`.
