+++ modules/howtos/pages/subdocument-operations.adoc
@@ -268,14 +268,14 @@ The `collection.lookupInAnyReplica()` method returns the first response -- from 
 
 [source,java]
 ----
-include::example$Couchbase.Examples.KV/SubDocument/Program.cs[tag=lookup-in-any-replica,indent=0]
+include::example$SubDocumentReplica.cs[tag=lookup-in-any-replica,indent=0]
 ----
 
 The `collection.LookupInAllReplicas()` fetches all available replicas (and the active copy), and returns all responses.
 
 [source,java]
 ----
-include::example$Couchbase.Examples.KV/SubDocument/Program.cs[tag=lookup-in-all-replicas,indent=0]
+include::example$SubDocumentReplica.cs[tag=lookup-in-all-replicas,indent=0]
 ----
 
 You may want to use `LookupInAllReplicas` to build a consensus,
